- some tweaks to MC version reading in the main Plugin class
Not going to go too much into version independence right now, but I wanted to at least get started with it (although I have a lot to do). The main goal at the moment is just to get ACP out for beta release, so that can of worms will come after.